Merged PR 210: Updated script to incorporate changes due to RSP-1507 Update Playwright Report
## What was the purpose of this ticket?

As a test engineer, I want to update the Playwright report to have more details regarding the test run, So that I can get detailed information regarding the completed test execution.

Acceptance Criteria

Update the 'Device Type' component in the report to display any one of the expected values as Desktop / iOS / Android as per the configuration of test run.
Add the 'Browser Name' component in the report and make sure it displays any one of the expected values as Google Chrome / Mozilla Firefox / Safari as per the configuration of test run.
Add the 'Browser Version' component in the report and make sure it displays the browser version where the tests are being executed.
Add the ‘Operating System' component in the report and make sure it displays the operating system 'Name & Version’ where the tests are being executed.
Add the 'Environment' component in the report and make sure it displays the value as 'QA' as per the configuration of test run.
Add the 'Execution Start Time' component in the report and make sure it displays the date and time when the test execution started.
7.Add the 'Execution End Time' component in the report and make sure it displays the date and time when the test execution ended.

## Jira Ref

Replace the `###` with the Jira Ticket number below

[RSP-1507](https://nihr.atlassian.net/browse/RSP-1507)

## Type of Change

Put [X] inside the [] to make the box ticked

- [] New feature
- [X] Refactoring
- [] Bug-fix
- [] DevOps
- [X] Testing

## Solution

What work was completed to cover off the story?

- List out changes made to the repo

**Modified file:-**

-/steps/automationtests/run_e2e_test.yaml

Updated script to incorporate changes due to RSP-1507 Update Playwright Report

Modified **(npx bddgen --tags "@authSetup or @Smoke" || true) && (npx playwright test || true) && (npx ts-node ./src/utils/Report.ts)** with
      **(npx bddgen --tags "@authSetup or @Smoke") && (npx playwright test)**

Modified **(npx bddgen --tags "@authSetup or @SystemTest or @IntegrationTest" || true) && (npx playwright test || true) && (npx ts-node ./src/utils/Report.ts)** with
      **(npx bddgen --tags "@authSetup or @SystemTest or @IntegrationTest") && (npx playwright test)**